Revert page_hash mutex performance patch
Summary:
Revert "Prevent leak of hash_mutex when page is read for dropped tablespace"

Revert "Small fix and additional asserts from buffer hash shard patch review."

Revert "Apply page hash mutex split patch from InnoDB"

This reverts the performance optimization that adds a separate mutex for the
buffer pool page hash.
It preserves the changes from other diffs while doing so. We are the
only user of the
page_hash change in 5.1. It exists in official InnoDB for 5.5 or 5.6.

The trigger for this is that I am trying to apply a patch to backport a fix for
mysql bug 61188
to fix slow drop innodb compressed table performance and that patch has a
conflict with the page_hash
changes in our branch. I have had several backports and merges to new 5.1
releases also conflict
with page_hash change. Conflict resolution takes more time that I want to spend
and I worry that the
code is correct afterwards. I don't think the performance improvement from this
patch is worth
the maintenance cost.

This is sysbench throughput in QPS for 8 to 1024 concurrent clients using fetch
one row by PK as
the query. The database is 2M rows in each of 8 tables and all data is cached.
I don't think there
is a significant perf impact from removing this diff.
clients               8    16    32    64   128   256   512  1024
with this diff:   25285 47007 93749 73760 64111 62889 61670 40565
before this diff: 25104 46847 92190 75735 64071 62832 61198 50633

Test Plan: mtr, sysbench
Reviewed By: ryandmack
Reviewers: nizamordulu, ryandmack, nponnekanti