Make admission control crash safe and simplify admission_control_enter
Summary:
This makes AC crash safe by setting an internal flag
(admission_control_disabled)
rather than assert when the AC state appears to be incorrect. Once disabled the
server must be
restarted to re-enable it. The SHOW STATUS variable Control_admission_disabled
displays the
value of the flag. An mtr test (admission_control_disabled.test) is added to
test error
handling.
"
This also simplifies admission_control_enter by removing the inner-most while
loop as that
was really redundant given the outer-most while loop. Performance is unchanged
with this edit
and the code is easier to read.
"
This changes admission_control_exit to check "qrp1 <= max_queries" where the
previous check
was "qr <= (max_queries + 1)". I think this check is more correct as another
thread should
be OK to run when the current count of running queries is less than the max,
and "<=" is
used in the check because qrp1 is the current count + 1.
"
Use int32 for queries_running and queries_waiting rather than int as the atomic
ops
macros use int32.
"
Change the arg in the call to store from TRUE (unsigned) to FALSE (signed) to
print
negative values should they occur.

Test Plan: mtr, sysbench

Reviewers: chip

Reviewed By: chip