Backport fix for mysql bug 54538
Summary:
This applies the patch for http://bugs.mysql.com/54538
Fix Bug #54538 - use of exclusive innodb dictionary lock limits performance.
This patch doesn't get rid of the need to acquire the dict_sys->mutex but
reduces the need to keep the mutex locked for the duration of the query
to fsp_get_available_space_in_free_extents() from ha_innobase::info().
To get this diff, do this on the 5.1 bzr trunk for mysql:
  bzr log -p 3351.61.2

Test Plan: mtr

Reviewers: nizamordulu

Reviewed By: nizamordulu