Reduce buffer pool mutex stalls during drop table.
Summary:
This is a workaround for http://bugs.mysql.com/51325. It adds the my.cnf
variable innodb_drop_table_phase1 that determines whether
buf_LRU_drop_page_hash_for_tablespace
is called during drop table processing. It was always called in the past to
reduce
the overhead of removing pages from the adaptive hash index. But it increases
the
time for which the buffer pool mutex is locked. The call isn't required as
buf_LRU_invalidate_tablespace repeats the work. This also adds a printf for
monitoring DROP TABLE perf which eventually might be removed. It starts with
"InnoDB: buf_LRU_invalidate_tablespace". This also adds two SHOW STATUS
counters
that measure the time for which the buffer pool mutex is locked during DROP
TABLE:
Innodb_drop_table_phase1_seconds - time in
buf_LRU_drop_page_hash_for_tablespace
Innodb_drop_table_phase2_seconds - time in buf_LRU_invalidate_tablespace after
    the call to buf_LRU_drop_page_hash_for_tablespace.
"
On a server with a 60G buffer pool and 3.2M pages used the buffer pool mutex is
locked for ~0.5 seconds (0.25 in buf_LRU_drop_page_hash_for_tablespace, and
0.25 in buf_LRU_invalidate_tablespace). So setting innodb_drop_table_phase1=OFF
should cut the mutex lock time from 0.5 to 0.25 seconds because the first
call is not made.
"
The "optimization" of calling buf_LRU_drop_page_hash_for_tablespace attempts
to reduce the overhead from removing pages from the hash index. It also removes
pages from the hash index and it can be expensive for
buf_LRU_invalidate_tablespace
to remove a page from the hash index because to do that it must release the
buffer pool mutex and then restart from the end of the LRU (a "rescan"). But,
fseg_free_extent -> btr_search_drop_page_hash_when_freed earlier in the DROP
TABLE processing removed all/most of the pages for this tablespace so the
rescan should be unlikely and the new monitoring lets us confirm that.

Test Plan: mtr

Reviewers: dmituzas

Reviewed By: dmituzas