Backport fix for bugs.mysql.com/59338 from 5.1.56
Summary:
This makes mysql write INSERT IGNORE and INSERT .. ON DUPLICATE KEY UPDATE to
the binlog even when no rows are changed to match what is done elsewhere and
what is
documented. I changed some tests from the original patch to avoid relying on so
many
new include files and to make the INSERT IGNORE test check that binlog events
are
written when no rows are inserted.

Test Plan: mtr

Reviewers: nizamordulu

Reviewed By: nizamordulu