Reserve empty space on pages of compressed tables to prevent compression failures
Summary:
Require empty space on the uncompressed pages of compressed tables in order to
prevent compression failures. The amount of data placed on a page is determined
by considering the N smallest pages that fail to compress and taking the median
of those page sizes.
The optimal amount of data to be placed on a page is determined when mysql
starts. There are two phases: gathering statistics and using gathered statistics
to restrict the amount of data on a page. In the first phase there is no
restriction on the amount of data that's placed on a page and the sizes of the
pages that failed to compress are stored in a binary tree. After the first phase
is complete the optimal amount is computed by taking the median (the root of the
tree) of the stored page sizes.
There are two variables: innodb_comp_fail_tree_size and
innodb_comp_fail_samples. Number of samples determine the number of compression
failures before we start restricting the data amount on a page in the first
phase. If innodb_comp_fail_tree_size is N then we store the N smallest sizes of
the pages that fail to compress during the first phase and the optimal page size
becomes the median of those N page sizes.

Test Plan: tpcc, linkbench benchmarks, run on shadows.

Reviewers: mcallaghan

Reviewed By: mcallaghan