Handle admission_control_enter failures that occur after query start
Summary:
When admision_control_enter call fails after query start the count of running
queries
is not incremented for the user so it should not be decremented by any calls to
admission_control_exit
that follow. If admission_control_enter fails at query start, then the exit
function is not called
This is only an issue for failures done when enter/exit are called for disk and
network IO. The AC
state is changed in the THD to prevent a decrement from being done in
admission_control_exit.
"
The debug macros are moved before a mutex is locked for processing of sleep().
The call to
admission_control_exit was also moved before the lock to avoid potential
deadlock in the future.
"
Changed admission_control.test to make it deterministic. Updated other tests
because SHOW PROCESSLIST
state is given now that the stall in sleep() is done earlier.

Test Plan: mtr

Reviewers: chip

Reviewed By: chip