Port innodb_fake_changes.patch from Percona
Summary:
This adds a session parameter, innodb_fake_changes, that prevents
rows from being locked and modified when set. It can be used to prefetch
transactions. When set COMMIT and auto-commit statements will return an
error. Non auto-commit statements are not foced to fail with this set.
Binlog events are not written for a session when this is set for changes
to InnoDB tables. All bets are off if a session mixes changes to MyISAM
and InnoDB.
"
The patch is at
http://bazaar.launchpad.net/~percona-dev/percona-server/5.1/view/head:/patches/innodb_fake_changes.patch

Test Plan: mtr

Reviewers: nizamordulu

Reviewed By: nizamordulu