Make the default for innodb_fake_changes_lock TRUE.
Summary:
When FALSE and no locks are taken another connection can delete the row
for which the fake-change connection is in the middle of doing secondary index
maintenance and not find the row and raise an error. row_upd_sec_index_entry is
one place where the error can occur. The error message is:
  InnoDB: error in sec index entry update in

Test Plan: mtr

Reviewers: nizamordulu

Reviewed By: nizamordulu