Don't wait on LOCK_log during ::unlog when committing transaction didn't make the binlog full
Summary:
This avoids stalls that occur because after writing to the binlog and releasing
LOCK_log the
transaction must lock LOCK_log again during ::unlog to check whether the binlog
is full. This creates
more dogpiles on LOCK_log especially when fsync is not fast via RAID write
cache.

This is QPS for sysbench with a cached database doing 1 auto-commit update per
query. Each
update statement updates 1 row (a non-indexed column). With fast fsync (RAID
accel on):
   8       16     32       64     128     256     512    1024   concurrent
clients
4385     4564    4583    4552    4512    4450    4210    4117   no group commit
4489     4626    4688    4665    4624    4540    4289    4220   no gorup commit
and this change
5625    10765   14745   20247   22456   22671   17403   11577   group commit
5712    11133   15778   21777   23791   23661   16016   12032   group commit
and this change
"
And with RAID accel off
 8      16      32      64     128     256     512    1024   concurrent clients
55      54      56      55      52      50      53      53      no group commit
53      54      54      54      54      48      51      50      no group commit
and this change
76     154     309     603    1076    2084    3423    3628      group commit,
gc timeout=10ms
84     151     312     646    1224    2267    3731    4396      group commit
and this change, gc timeout=10ms
85     190     316     605     984    1090    1084    1064      group commit
and this change, gc_timeout=1ms

Test Plan: mtr, sysbench with RAID accel on and off

Reviewers: nizamordulu

Reviewed By: nizamordulu