Add INNODB_BUFFER_POOL_PAGES to IS.table_statistics
Summary:
This column reports the number of pages in the InnoDB buffer pool for that
table. It only
counts pages on the LRU. Uncompressed pages for compressed tables are on the
unzip_LRU and support for
that will be added later.
"
The implementation trades complexity to reduce stalls on the buffer pool mutex.
That mutex is locked
when pages are added to and removed from the LRU. I don't want to lock another
mutex, lock up
the per-tablespace stats object and change the count when the buffer pool mutex
is locked. So the
count changes are deferred until the buffer pool mutex is not locked.

Test Plan: mtr

Reviewers: nizamordulu

Reviewed By: nizamordulu