Improve padding computation
Summary:
We introduce a new variable called innodb_comp_fail_max_fail_rate that
determines the maximum failure rate we are willing to tolerate for compressed
tables. We also drop innodb_comp_fail_threshold because usage of this variable
conflicts with the usage of comp_fail_max_fail_rate. Making both of them work
may make the code more complicated than it needs to be.

After innodb_comp_fail_samples compression failures takes place for an index,
we take the median page size of the smallest innodb_comp_fail_tree_size pages
that failed to compress. This becomes the maximum size of the data to be placed
on a page that belongs to this index.

Sometimes the failure rate doesn't become low enough for the max page size
determined after innodb_comp_fail_samples failures. If after
innodb_comp_fail_samples compression failures the failure rate is still greater
than innodb_comp_fail_max_fail_rate, we reset the internal counters for the
compression failures and compression operations, but we keep the list of the
smallest page sizes that failed to compress. This preserves the maximum page
size fixed so the future compression failures will make the median of the
comp_fail_tree smaller.

To give an example, let innnodb_comp_fail_max_fail_rate be 0.1 and let
innodb_comp_fail_samples be 200. Let the max page size (the median of the
smallest pages that failed to compress) after 200 comp failures be 15K. If at
the end of 200 compression failures the failure rate is greater than 0.1, then
we will continue to maintain the list of smallest pages that failed to compress.
If after 400 compression failures the failure rate goes below 0.1 and the max
page size becomes 14K then the final value for the max page size will be 14K and
we will no longer keep the list of smallest pages that failed to compress.

With this change we decrease the required minimum value of max page size from
UNIV_PAGE_SIZE/2 to UNIV_PAGE_SIZE/4. We initially picked UNIV_PAGE_SIZE/2 to
make sure that the the max data size on a page will not be less than the size of
the compressed page. We lower this for two reasons:
1- When fragmented, uncompressed tables do not use 100% of the page size so a
utilization of less than 50% can still save some space for compressed tables.
2- If the computed max page size is less than UNIV_PAGE_SIZE/2, allowing more
data on the pages of the index will just cause more compression failures without
any benefit. The data on the pages will still be less than UNIV_PAGE_SIZE/2. One
can ask whether such tables should be compressed. It is possible for a table
that some indexes on the table compress well and others don't but overall the
compressed size of the table is smaller than the uncompressed size. In such a
case it's possible for some of the indexes of the table to become inflated after
compression.

UNIV_PAGE_SIZE/4 is just a small enough random value. It assumes that the
inflation of the index of a compressed table is not more than 2X which I think
is the case for our UDB tables.

Test Plan:
mtr
shadows

Reviewers: mcallaghan

Reviewed By: mcallaghan