Fix bug 62100 -- fil_rename_tablespace times out and error is ignored
Summary:
This fixes http://bugs.mysql.com/62100. The problem is that
fil_rename_tablespace can timeout and return an error that is ignored.
The ignored error is not reported to the user. The timeout occurs when
there are pending writes for the renamed tablespace but the writer
threads are not started. They are not started because the doublewrite
buffer is not full and any attempt to request more writes and make
it full are blocked by the stop_ios check in
fil_mutex_enter_and_prepare_for_io. The stop_ios check is boosted
before the check for too-many open file descriptors to make it
deterministic.

I modified fil0fil.c and then ran a reproduction case to confirm
this reproduces and then fixes the problem. It is not possible
to write an mtr test to reproduce this without such a hack as
the race is hard to reproduce otherwise.

Test Plan: mtr, custom

Reviewers: nizamordulu

Reviewed By: nizamordulu