Fix btr_cur_optimistic_update() for compressed tables
Summary:
btr_cur_optimistic_update() tries to execute an update query without splitting
a page. In order to do this, it has to make sure that the update can be
performed without splitting the page. This means that the page must have enough
space to place the updated record after the old record is deleted. The max size
for the updated record used to be computed via the following formula:

  max_size = old_rec_size + page_get_max_insert_size_after_reorganize(page, 1);

This assumes that page can be reorganized before the updated record is placed.
The problem is that reorganized page has different compressibility
characteristics than the fragmented page. In order to fix this issue, I changed
this to:

  max_size = old_rec_size + (page_zip ? page_get_max_insert_size(page, 1) :
page_get_max_insert_size_after_reorganize(page, 1));

So that the max_size for the updated record does not depend on the free space
on the reorganized page. This is still wrong, however, because the emptied space
from the delted record (old_rec_size) can not be used for the new record unless
the page is reorganized. Therefore the correct computation must be as follows:

  max_size = page_zip ? page_get_max_insert_size(page, 1) : (old_rec_size +
page_get_max_insert_size_after_reorganize(page, 1));

Test Plan:
Test it on the crashed compressed slaves.
mtr

Reviewers: mcallaghan

Reviewed By: mcallaghan