Signal the IO thread when rpl_transaction_enabled is on and logs are purged
Summary:
This was found while reading code for bugs.mysql.com/64503. There
is code in purge_first_log that resets ignore_log_space_limit and signals
the IO thread. This code is not used when rpl_transaction_enabled is ON.
This diff depends on the previous diff for this task/bug.

Test Plan: mtr

Reviewers: nizamordulu

Reviewed By: nizamordulu