Make Segment Fill Factor a Dynamic Variable.
Summary:
segment_reserve_factor determines the ratio of empty pages, the last segment in an ibd can have before innodb creates a new extent. If a new page is needed and the last segment has less than segment_reserve_factor * (current number of pages in this segment) empty pages, then a new extent is created and the new page is allocated from the new extent. The empty pages are useful when B-tree grows so that pages in the upper level of the B-tree can be allocated contigiously. An extent is 64 pages, and a segment is 256 extents.

Before this diff, this variable was not configurable and it was determined by 1/x where x was the value defined by the compile time define FSEG_FILLFACTOR. This used to be 0.125 i.e. if fewer than 1/8th pages were empty then InnoDB would allocate a new extent leaving many empty pages in the current extent. This caused lots of fragmentation.

The ideal value for this variable needs investigation. The current default will have the previous behavior. A segment has 16384 pages, so for eg. if we are ok with leaving at most 16 empty pages per segment, then we can set innodb_segment_reserve_factor=0.001.

innodb_segment_reserve_factor.test loads data into two identical tables but sets a lower segment_reserve_factor before loading data to the second table. This makes the second table about 10% smaller than the first table.

The default value for this variable is set to 0.01 which should mean 1% of the pages will be allocated for growth.

Test Plan:
innodb_segment_reserve_factor.test
mtr
slaves: I ran two slaves one with segment_reserve_factor=0.001 and other with 0.125 (current default value). Both servers were compressed. Statistics below after fragmentation:

  table                                    factor=0.125        factor=0.001
                                           size(empty page%)   size(empty page%)
  db8526.assoc_info                        7.8G(14.1%)         6.8G(2.4%)
  db8526.fbid                              6.7G(4.3%)          6.5G(1.9%)
  db8526.minifeed_stories                  8.9G(14.2%)         7.7G(2.0%)
  db8526.assoc_friends                     1.9G(14.1%)         1.7G(2.4%)
  db8526                                   109G                103G
  total(db8526,db8106,db8806,db9226)       456G                431G

shadows: We need to figure out the ramifications (possibly io-related) of this change and find the ideal value for innodb_fsegment_reserve_factor. I prefer to do the shadow testing along with the other compression changes. I don't think page allocation depends too much on concurrency and that's the only difference between a pure-slave and a shadow in terms of page allocation.

Reviewers: mcallaghan

Reviewed By: mcallaghan