MySQL discards fractional part of the default values of system variables
Summary:
The struct used to represent a global variable option my_option in my_getopt.h, has a field of type longlong to represent the value for scalar variables. This truncates the double values to integer if the global variable is not defined in my.cnf.

This currently affects us. innodb_comp_fail_max_fail_rate's default value is 0.1 but it is set to 0 in production because of this bug.

Test Plan:
1- run mysql check the default value for innodb_comp_fail_max_fail_rate.
2- specify innodb_comp_fail_max_fail_rate in my.cnf, check value.
3- specify innnodb_comp_fail_max_fail_rate in command line, check value.
4- check other global vars make sure this doesn't affect them.
5- check the output of mysqld --help.

Reviewers: mcallaghan

Reviewed By: mcallaghan