In buf_LRU_free_block, update the checksum if the corresponding sys var is set.
Summary:
in buf_LRU_free_block() the checksum of a compressed page is updated even if the page is dirty. For now,
we allow this redundancy if innodb-extra-checksums-unzip-lru is ON. We can remove this after we are sure that
this doesn't break anything.

Test Plan:
mtr
slaves
shadows
following should happen in order to properly test this:

- trigger the eviction for the uncompressed frame of a compressed clean page. see what happens during flush.
- trigger the eviction for the uncompressed frame of a compressed dirty page. do not modify the page after the call to buf_LRU_free_block. see what happens during flush.
- trigger the eviction for the uncompressed frame of a compressed dirty page. modify the page after the call to buf_LRU_free_block(). see what happens during flush.

Reviewers: mcallaghan

Reviewed By: mcallaghan