Add more zlib options as system vars. (REQUIRES CLEAN BUILD FOR STRESS TESTS)
Summary:

This diff exposes more zlib config parameters as system vars. The most
important of this variables is innodb_zlib_wrap which makes zlib not
write adler32 checksums between compressed blocks. The default value for
wrapping is OFF now. This should save some calls to adler32 and some
space.

Changes are backwards compatible. In other words it's ok to copy an old
compressed server using xtrabackup to a server with the rpm that has
this change. This is because I updated the transaction log records to
include the value for wrap and the compression strategy in addition to
compression level. This was fairly straightforward because I did the
same earlier with compression level. Prior to this diff, we stored the
compression level in a byte. I used the spare bits in that byte to store
values for the new options.

innodb_zlib_wrap and innodb_zlib_strategy also effect blob compression.

it's ok to change these values dynamically. this won't break xtrabackup
or crash recovery.

I needed to backport inflateReset2() from a future version of zlib in
order to reset a decompression stream without allocating memory for its
internal state and with specifying windowBits. This was necessary to
automatically determine whether a compressed page was compressed using
adler32 checksums or not.

There should not be any issues in terms of backwards compatibility with
this diff.

After this diff, a clean build is required for stress tests to pass.

Test Plan:
mtr

xtrabackup: with the change in compressed page size we need to make sure
that crash recovery works fine. (compressed page isize will be smaller
because of the decrease in number of adler32s.)

mtr tests that change these options and run queries similar to
innodb_change_compression_level.

Reviewers: mcallaghan, dmituzas

Reviewed By: mcallaghan