Make mysqldump not lock all tables when --single-transaction is used
Summary:
Change the behavior of --single-transaction to use

START TRANSACTION WITH CONSISTENT INNODB SNAPSHOT

to begin transactions, which returns the binlog file and offset.  This
is later used if --master-data is specified and becomes part of the
output (otherwise it just avoids the lock tables).

Test Plan: run mysqldump, verify it is executing the commands expected

Reviewers: nponnekanti, hfisk

Reviewed By: hfisk