fix test regression of "use seperate threads for tcp/ip socket and unix socket."
Summary:
"use seperate threads for tcp/ip socket and unix socket." caused on test
regression. It is not a real bug, and the problem is it produce one
'Error' in the test output file, which then cause the test fail.

Test Plan:
run 51test on failed testcases mcallaghan sent to me.
innodb_plugin.innodb_all_direct still fails, but i think
it is not related to this changes.

Reviewers: mcallaghan, dmituzas, chip

Reviewed By: mcallaghan