Allow more time to restart in tests in 5.1.
Summary:
Debug builds are slow, so some big tests time-out on restart.

This changes the hard-coded test restart timeout to roughly 5 minutes,
instead of the previous value of roughly 50 seconds.

Test Plan:
This change was made in 5.6, and it has fixed this issue so far.
Jenkins' nighly tests will show if this fixes the occasional fail

Reviewers: nizamordulu

Reviewed By: nizamordulu