Don't free buffers not owned by String
Summary:
http://bugs.mysql.com/bug.php?id=65562

String::shrink was an accident waiting to happen. The call stack to
the accident is:
  String::shrink - called because my_b_read call fails
  String::append(IO_CACHE* file, uint32 arg_length)
  Log_event::read_log_event
In the call stack above the my_b_read call fails because the end of the binlog
is corrupt on crash recovery. Newer versions of MySQL truncate the end of the
binlog when it is corrupt after a crash.

An earlier change by me to add rpl_event_buffer_size makes this possible because
with that change, the instance of String named "packet" in mysql_binlog_send
provides a buffer to "packet", thus "packet" doesn't own the buffer. But in the
call stack above it will attempt to free it.

Test Plan: mtr

Reviewers: steaphan

Reviewed By: steaphan