Make innodb stress tests use less global state
Summary:
Change the tests to use a tmp directory from MYSQL_TMP_DIR
rather than /tmp. This will help when tests run concurrently on the same host.

Disabled rpl_peak_lag, it is not deterministic and I am willing to accept the
risk of not testing that the reporting is accurate. Task #1113343 filed to fix it.

Test Plan: mtr

Reviewers: steaphan

Reviewed By: steaphan