Back-port fix for authentication bug (#64884).
Summary:
They were returning the result of memcmp as a my_bool to see if a
password was valid or not.

Now uses the test() macro from include/my_global.h to make it 0 or 1.
This is the exact fix from upstream, so it should merge cleanly.

Test Plan: Jenkins (re-enabling auth test).

Reviewers: mcallaghan

Reviewed By: mcallaghan