Replication IO thread retries more frequently
Summary:
Change get_master_version_and_clock so that it returns a
retryable error code when queries to the master fail, regardless
of the failure. It will still return a non-retryable error
if the master query succeeds but returns data the slave views
as bad.

Test Plan: mysqltoolbox.sh 51tests

Reviewers: mcallaghan

Reviewed By: mcallaghan