Optimize prefix index queries to skip cluster index lookup when possible
Summary:
Currently InnoDB will always fetch the clustered index (primary
key index) for all prefix columns in an index, even when the value of a
particular record is smaller than the prefix length.  This change
optimizes that case to use the record from the secondary index and avoid
the extra lookup.

Test Plan: mtr

Reviewers: nizamordulu, jtolmer

Reviewed By: nizamordulu