Fix unstable test rpl.rpl_event_buffer_size_bug
Summary:
There was a race in the test. The IO thread was starting, hitting
an error and exiting faster than include/start_slave.inc could
detect that it ever started and so the test would timeout waiting
for the thread to start up.

Test Plan: mtr rpl.rpl_event_buffer_size_bug

Reviewers: steaphan

Reviewed By: steaphan