Increased group commit hang times.
Summary:
With debug=full, the times used for group commit exceptions were too low.

These were:
	group_commit_hang_disable_secs = 60;
	group_commit_hang_log_secs = 2;

So, I multiplied them by 30, but only when built with ENABLED_DEBUG_SYNC.

I also multiplied the value specified in one test by 30,
which is fine since it is a debug-only test.

I also replaced the values with # in another test,
since it shows the values, but doesn't care what they are.

Test Plan: Jenkins found this.

Reviewers: chip, jtolmer

Reviewed By: jtolmer