Further crash proof rpl_transaction_enabled
Summary:
Previous work left a crash-recovery case which would cause a slave to
stop replicating; if the slave crashed after a transaction was written
to the bin log but before it was committed to InnoDB. This diff fixes
the code to handle that case by storing 2 sets of replication
information in the InnoDB system header, one each for the last prepared
and committed transactions. When doing recovery determine which to use
based on whether crash recovery rolled back a prepared transaction.

Test Plan:
mtr
repeatedly crash a shadow

Reviewers: mcallaghan, chip

Reviewed By: mcallaghan