Make padding_rounds_successful_max a dynamic variable and change the default value to 2.
Summary:
This revision changes the default value for the maximum value for the successful
rounds for the linear padding algorithm. I am making this change based on my
observations on the production instances that had bad paddings but their padding
didn't go down fast enough. I will test this on shadow instances before the rpm gets deployed.
I am also making it a dynamic variable so that we can change it if we need.
Also change the default values for the following:
innodb_padding_max from 0.5 to 0.4: this will disallow the padding value to grow more than 0.4*UNIV_PAGE_SIZE.
In the last rpm the default value for this was .75 which was too high for our instances. I looked at
our instances and except for the fbid table, none of the tables needed more than 6K padding.
innodb_padding_max_fail_rate from 0.05 to 0.1: 5% is too low for our production we can afford to have 10% fail
rate and that would improve compression rate.

Test Plan: OK

Reviewers: zshao

Reviewed By: zshao