Use a separate mutex and hash table for per-tablespace stats
Summary:
This is done to reduce mutex contention on fil_system->mutex as that
mutex is briefly locked before every page read and we don't want to make those
operations wait. The new hash is fil_system->stats_hash and it uses an array of
mutexes. Entries are searched by tablespace ID. There is a small window right
after a tablespace is created and right before it is deleted when there might not
be an entry in stats_hash for a valid tablespace. The hash can be searched via the
function fil_stats_get_by_id.

Test Plan:
mtr, sysbench TBD

Reviewers: steaphan, nizamordulu

Reviewed By: steaphan