Fix logic for when to get a free block
Summary:
This code was changed by me incorrectly at the end of the code review
process when I first committed it. It was correct until that last-minute change.
The last minute change moved it into the else branch that is infrequently run,
so the performance benefit doesn't occur. Note that it is incorrect for
performance, but not for correctness.

Test Plan:
mtr

Reviewers: steaphan

Reviewed By: steaphan