Attempt to fix intermittent failure in rpl.rpl_slave_load_tmpdir_not_exist
Summary:
The test failed due to an unexpected warning in the err log from the
slave IO thread, "[Warning] Slave I/O: Get master SERVER_ID failed with
error: Lost connection to MySQL server during query, Error_code: 2013."

My best guess of the cause is that there was a race in the test between
the IO thread starting, which is asynchronous, and the test trying to
shut it down and run RESET SLAVE. Most tests don't have this race because
they use include/start_slave.inc which waits for both slave threads to
start before returning control to the test. This test can't use that
include because it expects an error in the SQL thread.

Test Plan: Jenkins

Reviewers: steaphan

Reviewed By: steaphan