[Handle multiple files being streamed to mysqlbinlog]
Summary:
When multiple files are streamed, the binlog magic number
occurs in the middle of stream and is causing problems.
Skip such binlog magic magic numbers that occur in the middle
of stream.
Also, do not allow --stop-position with stream input because
we do not know when the last file starts.

Test Plan:
Added tests that stream multiple binlog files

Reviewers: jtolmer

Reviewed By: jtolmer