[rpl.rpl_row_mysqlbinlog is using --stop-position with stdin]
Summary:
As part of "[Handle multiple files being streamed to mysqlbinlog]", we
disallowed stop-position with stdin. This is because when multiple
binlog files are being streamed in stdin (more often than not we have
more than 1 binlog file), it gets applied to first file contents when
the caller may actually be expecting that it gets applied to last file
(remember that it gets applied to last file when input is a list of
regular files.)

rpl.rpl_row_mysqlbinlog has 2 tests using --stop-position with stdin.
Change them to expect error.

Test Plan: Revert Plan: OK

Reviewers: jtolmer

Reviewed By: jtolmer