Fix mysqlbinlog in reading big records from stdin
Summary:
When mysqlbinlog tries to read a big record (over 128KB) from stdin,  the glibc read in my_read will return 128KB instead of the full size.   This causes _my_b_read to fail.

There are 2 solutions:
A. In mysqlbinlog.cc, enable MY_FULL_IO, and in my_read.c, fix the MY_FULL_IO logic.  It seems MY_FULL_IO was never used, so I am not surprised that the logic has a bug (when dealing with EOF).

B. In _my_b_read, add a loop to the call to my_read.

This diff implements A which I prefer because that seems what MY_FULL_IO was defined for.

Test Plan:
Tried manually against a real binlog:
$ cat binlog | mysqlbinlog -

Will need to add a test to mtr.

Reviewers: nponnekanti, jtolmer, mcallaghan

Reviewed By: nponnekanti