Compile-time option to increase PAGE_ZIP_DIR_SLOT_SIZE to 3 bytes
Summary:
Preparation work for 32KB page sizes.

The dense directory of a compressed page stores the following for each
record of the page:
  * Whether the record is owned
  * Whether the record is deleted
  * The offset of the record on the page

14 bits could store an offset large enough to cover a 16KB page, and so
a dense directory entry fit in 2 bytes. 14 bits is not large enough for
an offset into a 32KB page so change it to be 3 bytes, 2 bytes for the
offset and 1 byte for the other info.

That change also required changes to:
  * PAGE_ZIP_DIR_SLOT_MASK
  * PAGE_ZIP_DIR_SLOT_OWNED
  * PAGE_ZIP_DIR_SLOT_DEL
  * mach_read_from_3/mach_write_to_3 instead of the 2 byte versions
    when writing the entries in the dense directory
  * Switch the shifts for PAGE_ZIP_DIR_SLOT_DEL and PAGE_ZIP_DIR_SLOT_OWNED
    from 8 to 16

Test Plan: mtr

Reviewers: nizamordulu

Reviewed By: nizamordulu