Make padding linear increment dynamic
Summary:
Make the padding linear increment dynamic so that tests
can set it higher in order to see the benefit with smaller
data sets.

Also changes the padding variables to require an argument
since it doesn't make sense to leave the argument off when
specifying them on the command line / in my.cnf.

Test Plan: mtr

Reviewers: nizamordulu

Reviewed By: nizamordulu