Revert "remove the unnecessary calls to btr_cur_optimistic_insert()"
Summary:
This reverts an earlier commit, and merges the related changes from
Oracle regarding btr_cur_pessimistic_insert() and
btr_cur_optimistic_insert().

That was an attempt to reduce the number of compression failures
that happen while the B-tree latch is locked. This was one of the causes
of mutex contention for workloads on compressed tables, and it happened
as follows:

1- row_ins_index_entry() is called during the insertion of a record.
2- InnoDB attempts to insert a record to a page while holding the page
latch only (search for BTR_MODIFY_LEAF in code). During this operation,
record gets inserted to the uncompressed page but page needs to be
recompressed and recompression of the page fails.
3- Because 2 failed, InnoDB retries the insert operation, now holding
the b-tree latch so that it can modify the b-tree to make room for the
insert. During this phase InnoDB still first attempts to insert the record
to the page, and it causes the same compression failure that it did in 2.
This compression failure is costly because it's happening while b-tree latch
is locked.

And fixed the redundant compression failures by eliminating the calls to
btr_cur_optimistic_insert() during the phase 3 above. This caused a
problem because between the phases 2 and 3, the page to which the record
is inserted may change and it may become available for the record
insertion.

Note that this diff brings back the mutex contention problem caused by
compression failures. I will address that in a separate diff.

Test Plan: mtr

Reviewers: jtolmer

Reviewed By: jtolmer